Experimental: load positions of idents through shadowed module aliasing
This allows, for example, to jump to the definition of LibIndex.Misc.unique_subdir, while
we would previously only get the shadowing signature from libIndex.mli

It should be made so that loading the cmt isn't done once for every position looked up though
(not a problem atm since we usually just query one loc).

Only works on >4.02, obviously.